[REFACTOR#175] 개발용 회원가입, 로그인 API 별도 분리 및 Profile 지정 (Local, Dev)
