+++ src/main/java/com/togedy/togedy_server_v2/domain/user/api/UserController.java
@@ -1,8 +1,6 @@
 package com.togedy.togedy_server_v2.domain.user.api;
 
 import com.togedy.togedy_server_v2.domain.user.application.UserService;
-import com.togedy.togedy_server_v2.domain.user.dto.PatchUserOnboardingRequest;
-import com.togedy.togedy_server_v2.domain.user.dto.CreateUserRequest;
 import com.togedy.togedy_server_v2.domain.user.dto.GetMyPageResponse;
 import com.togedy.togedy_server_v2.domain.user.dto.GetMySettingsResponse;
 import com.togedy.togedy_server_v2.domain.user.dto.GetMyStatusResponse;
@@ -11,13 +9,12 @@ import com.togedy.togedy_server_v2.domain.user.dto.GetNicknameValidationResponse
 import com.togedy.togedy_server_v2.domain.user.dto.PatchMarketingConsentedSettingRequest;
 import com.togedy.togedy_server_v2.domain.user.dto.PatchProfileRequest;
 import com.togedy.togedy_server_v2.domain.user.dto.PatchPushNotificationSettingRequest;
+import com.togedy.togedy_server_v2.domain.user.dto.PatchUserOnboardingRequest;
 import com.togedy.togedy_server_v2.global.response.ApiResponse;
 import com.togedy.togedy_server_v2.global.security.AuthUser;
 import com.togedy.togedy_server_v2.global.util.ApiUtil;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
-import java.util.HashMap;
-import java.util.Map;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.MediaType;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
@@ -26,7 +23,6 @@ import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PatchMapping;
-import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
@@ -39,21 +35,6 @@ import org.springframework.web.bind.annotation.RestController;
 public class UserController {
 
     private final UserService userService;
-
-    @Operation(summary = "간편 회원가입", description = """
-            
-            이메일과 닉네임을 기반으로 회원가입을 한다.
-            
-            """)
-    @PostMapping("/sign-up")
-    public ApiResponse<Map<String, Long>> createUser(@Validated @RequestBody CreateUserRequest request) {
-        Long userId = userService.generateUser(request);
-
-        Map<String, Long> response = new HashMap<>();
-        response.put("userId", userId);
-
-        return ApiUtil.success(response);
-    }
 
     @Operation(summary = "닉네임 검증", description = "온보딩 전에 닉네임의 공백, 길이, 비속어, 중복 여부를 검증한다.")
     @GetMapping("/nickname/validate")
